requirements.txt
DateTime==4.3
netmiko==3.4.0
python-dotenv==0.15.0

